+++ src/configure.ts
@@ -17,17 +17,23 @@ export function jestPreviewConfigure(
   options.externalCss.forEach((cssFile) => {
     const basename = path.basename(cssFile);
     // Avoid name collision
-    const destinationBasename = `jp-${basename}`;
+    // Add `global` to let `jest-preview` server that we want to cache those files
+    const destinationBasename = `cache-${basename}`;
     const destinationFile = path.join(
       './node_modules/.cache/jest-preview-dom',
       destinationBasename,
     );
     // TODO: To move to load file directly instead of cloning them to `.cache`
     // Move together with transform
-    if (!fs.existsSync(destinationFile)) {
-      fs.copyFile(cssFile, destinationFile, (err: any) => {
-        if (err) throw err;
-      });
-    }
+    // TODO: To cache those files. We cannot cache them by checking if files exists
+    // Since content of the files might changes and it won't be copied over
+    // Can we send a websocket event to preview server and let server remember location of the files in the memory?
+    // That way, we can don't have to copy files to disk
+    // Memory is faster than disk anyway!!!!
+    // if (!fs.existsSync(destinationFile)) {
+    fs.copyFile(cssFile, destinationFile, (err: any) => {
+      if (err) throw err;
+    });
+    // }
   });
 }