+++ src/index.ts
@@ -1,6 +1,8 @@
 import { processFile, processFileCRA, processCss } from './transform';
-import { preview } from './preview';
+import { debug } from './preview';
 import { jestPreviewConfigure } from './configure';
 
 export { processCss, processFile, processFileCRA, jestPreviewConfigure };
-export default preview;
+export default {
+  debug,
+};