+++ src/preview.ts
@@ -1,6 +1,6 @@
 import fs from 'fs';
 
-export function preview(element: Element): void {
+export function debug(element: Element = document.body): void {
   if (!fs.existsSync('./node_modules/.cache/jest-preview-dom')) {
     fs.mkdirSync('./node_modules/.cache/jest-preview-dom', {
       recursive: true,
@@ -10,6 +10,7 @@ export function preview(element: Element): void {
   import('./styled-components').then(({ getStyle }) => {
     // If user use styled-components
     if (getStyle) {
+      // TODO: We can send this data via websocket instead of writing to disk
       fs.writeFileSync(
         './node_modules/.cache/jest-preview-dom/jp-styled-components.css',
         getStyle(),